requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
gunicorn>=22.0.0
google-cloud-firestore>=2.17.0
google-cloud-storage>=2.18.0
jinja2>=3.1.4
python-multipart>=0.0.9
python-dotenv>=1.0.1
pydantic>=2.10.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4